[FEAT] 공지사항 삭제, 수정기능 추가 개발

+++ app/src/main/java/org/sopt/app/interfaces/notice/NoticeRepository.java
@@ -3,5 +3,5 @@ package org.sopt.app.interfaces.notice;
 import org.sopt.app.domain.entity.Notice;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface NoticeRepository extends JpaRepository<Notice, Integer> {
+public interface NoticeRepository extends JpaRepository<Notice, Long> {
 }
